move swagger auto-generated code to internal, export sub clients

+++ scripts/swagger-gen.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
-
 SWAGGER_IMAGE="quay.io/goswagger/swagger"
-API_VERSION="v1.10.0"
-SWAGGER_FILE="https://raw.githubusercontent.com/goharbor/harbor/${API_VERSION}/api/harbor/swagger.yaml"
-
+API_VERSION_V1="v1.10.0"
+API_VERSION_V2="v2.0.0"
+SWAGGER_FILE_V1="https://raw.githubusercontent.com/goharbor/harbor/${API_VERSION_V1}/api/harbor/swagger.yaml"
+SWAGGER_FILE_V2=https://raw.githubusercontent.com/goharbor/harbor/${API_VERSION_V2}/api/${API_VERSION_V2%.0}/swagger.yaml
 # go-swaggers documentation on swagger operations at the moment is really sparse,
 # so here is a bit of explanation from code observations.
 # go-swagger accepts operation names from command line "--operation" flag, to filter which operations to generate.
@@ -62,11 +62,28 @@ for i in "${swagger_operations[@]}"; do
   operation_flags+="--operation=${i} "
 done
 
-docker run --rm -e GOPATH="${HOME}/go:/go" -v "${HOME}:${HOME}" -w "$(pwd)" ${SWAGGER_IMAGE} \
+if [[ "$1" = "v1" ]]; then
+  echo "using the v1 swagger file (${API_VERSION_V1})"
+  docker run --rm -e GOPATH="${HOME}/go:/go" -v "${HOME}:${HOME}" -w "$(pwd)" ${SWAGGER_IMAGE} \
   generate client \
   --skip-validation \
-  --model-package="api/${API_VERSION}/model" \
+  --model-package="api/${API_VERSION_V1}/model" \
   --name="harbor" \
-  --client-package="api/${API_VERSION}/client" \
-  --spec="${SWAGGER_FILE}" \
+  --client-package="api/${API_VERSION_V1}/client" \
+  --spec="${SWAGGER_FILE_V1}" \
   ${operation_flags}
+fi
+
+if [[ "$1" = "v2" ]]; then
+  set -x
+  echo "using the v2 swagger file (${API_VERSION_V2})"
+    docker run --rm -e GOPATH="${HOME}/go:/go" -v "${HOME}:${HOME}" -w "$(pwd)" ${SWAGGER_IMAGE} \
+  generate client \
+  --skip-validation \
+  --model-package="api/${API_VERSION_V2%.0}/model" \
+  --name="harbor" \
+  --client-package="api/${API_VERSION_V2%.0}/client" \
+  --spec="${SWAGGER_FILE_V2}"
+fi
+
+
